dev-requirements.txt
django-debug-toolbar==0.8.3
django-template-repl==0.3.0
yolk==0.4.1
